+++ application/views/__css.php
@@ -1,0 +1,10 @@
+<?php
+defined('BASEPATH') OR exit('No direct script access allowed');
+/**
+ * Load css styling files (in the head section)
+ */
+?>
+<!-- Load files for CSS Styling -->
+{styles}
+<link href="{link}" rel="stylesheet" type="text/css"/>{/styles}
+<!-- End Load files for CSS Styling -->