+++ application/views/__js.php
@@ -1,0 +1,10 @@
+<?php
+defined('BASEPATH') OR exit('No direct script access allowed');
+/**
+ * Load javaScript files (at the end)
+ */
+?>
+<!-- Load files for JS scripting -->
+{scripts}
+<script src="{link}" type="text/javascript"></script>{/scripts}
+<!-- End Load files for JS scripting -->