+++ application/views/_searchResult.php
@@ -1,0 +1,28 @@
+<table id="searchResult" class="display">
+	<thead>
+		<tr>
+			<th></th>
+			<th>Course Code</th>
+			<th>Booking Type</th>
+			<th>Instructor</th>
+			<th>Day of Week</th>
+			<th>Start Time</th>
+			<th>End Time</th>
+
+		</tr>
+	</thead>
+	<tbody>
+		{results}
+		<tr>
+			<td>{type}</td>
+			<td>{courseCode}</td>
+			<td>{courseType}</td>
+			<td>{instructor}</td>
+			<td>{day}</td>
+			<td>{periodStart}</td>
+			<td>{periodEnd}</td>
+
+		</tr>{/results}
+	</tbody>
+</table>
+